Deleted 2D_resnet colab notebook
This notebook file contained code that dealt with experimentation of using a resnet architecture with 2D convolutional layers instead of 1D. In the end 1D convolutions proved to be more intuitive for the task at hand and this file was not further developed. Therefore, it does not hold any relevant information for the final product.